requirements.txt

flask
pymongo
requests
python-dotenv
